pyproject.toml
[project]
name = "sphinx-terraform-ch"
version = "2026.04.28.2"
description = "A fork of the sphinx-terraform extension."
readme = "README.md"
requires-python = ">=3.11"
license-files = ["LICENSE"]
license="BSD-2-Clause-Patent"
dependencies = [
    "docutils>=0.22.4",
    "importlib-metadata",
    "sphinx>=8.0.0",
    "myst-parser[linkify]>=5.0.0",
    "pygohcl>=1.2.7",
    "sphinx-argparse>=0.5.2",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
tfreport = "tfreport:main"
tfmapexpand = "tfmapexpand:main"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name="sphinx_terraform_ch"
module-root="src"

[dependency-groups]
dev = [
    "build>=1.4.4",
    "coverage>=7.13.5",
    "pytest>=9.0.3",
]
